refactor(workflow): extract DAG inspection models below size budget
